+++ frontend/src/App.tsx
@@ -1,7 +1,21 @@
 import { memo } from 'react';
+import { createGlobalStyle } from 'styled-components';
+import Router from 'Router';
+
+const GlobalStyle = createGlobalStyle`
+	body {
+		color: white;
+		background-color: #596F99;
+	}
+`;
 
 function App() {
-	return <div className='App'>HELLO WORLD</div>;
+	return (
+		<>
+			<GlobalStyle />
+			<Router />
+		</>
+	);
 }
 
 export default memo(App);